pyproject.toml
[project]
name = "fastapi-user-items-template"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Patrick Saint-Denis", email = "patricksaintdenis88@gmail.com"},
]
dependencies = [
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.31",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.2",
    "fastapi>=0.111.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "python-decouple>=3.8",
    "faker>=26.0.0",
    "pytest-cov>=5.0.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.2.2",
    "pytest-alembic>=0.11.0",
    "docker>=7.1.0",
]
